requirements.txt
biopython==1.85
numpy==2.2.3
pandas==2.2.3
scikit_learn==1.6.1
setuptools==53.0.0
